refactor : network 모듈 모델명 변경
