+++ network/src/main/java/com/kky/pokedex/network/model/PokemonResponse.kt
@@ -2,13 +2,19 @@ package com.kky.pokedex.network.model
 
 import com.google.gson.annotations.SerializedName
 
-data class Pokemon(
-    val id: Int,
-    val names: List<PokemonName>,
-    @SerializedName("flavor_text_entries") val descriptions: List<FlavorText>
+data class PokemonResponse(
+    val id: Long,
+    val stats: List<StatInfo>,
+    val abilities: List<AbilityInfo>,
+    val weight: Int,
+    val height: Int
 )
 
-data class PokemonName(
-    val name: String,
-    val language: Info
+data class StatInfo(
+    @SerializedName("base_stat") val baseStat:Int,
+    @SerializedName("stat") val stat: Info
+)
+
+data class AbilityInfo(
+    val ability: Info
 )
